Charizard-Drogon-WebServer/backend-python/RPA/tarefas_ONS/requirements.txt
streamlit
pandas
tabula-py
PyPDF2
python-dotenv
google-generativeai
openpyxl
pytesseract
pdf2image
Pillow
camelot-py[base]